+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Adds prefab values for `java.util.LinkedHashSet`. ([Issue 1231](https://github.com/jqno/equalsverifier/issues/1231))
+
 ## [4.5] - 2026-04-17
 
 ### Added